# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ORVELL PULSE Wholesale ERP Backend API",
        "description": "Production REST API documentation for ORVELL PULSE Wholesale ERP system with Spatie RBAC, multi-company isolation, inventory tracking, cashier settlement, and automated reporting.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Default API Host"
        }
    ],
    "paths": {
...
        "/api/dashboard/overview": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Retrieve live operational KPIs, cash in till, inventory breakdown, and alerts for company",
                "description": "Get Real-time Dashboard Overview",
                "operationId": "bb0b4c093578c1503e89563687f588e1",
                "responses": {
                    "200": {
                        "description": "Live dashboard overview retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "company": {
                                                    "type": "object"
                                                },
                                                "financials": {
                                                    "type": "object"
                                                },
                                                "inventory": {
                                                    "type": "object"
                                                },
                                                "orders": {
                                                    "type": "object"
                                                },
                                                "low_stock_alerts": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "recent_activity": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/reconciliation/eod/generate": {
            "post": {
                "tags": [
                    "EOD Reconciliation"
                ],
                "summary": "Generate End-of-Day financial and inventory reconciliation snapshot",
                "description": "Generate EOD Reconciliation Report (Admin / Manager Only)",
                "operationId": "571926306d3b335660e68b8374c52ed5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "date"
                                ],
                                "properties": {
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-08-09"
                                    },
                                    "company_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "EOD reconciliation generated successfully"
                    },
                    "403": {
                        "description": "Unauthorized. Only Admins and Managers can generate EOD reconciliation reports"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/reconciliation/eod/show": {
            "get": {
                "tags": [
                    "EOD Reconciliation"
                ],
                "summary": "Retrieve existing End-of-Day reconciliation report by date",
                "description": "Show existing EOD Reconciliation Report",
                "operationId": "decc04566c1bd4a875fa79a3dbba6802",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Reconciliation report date (YYYY-MM-DD)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-08-09"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "EOD report retrieved successfully"
                    },
                    "404": {
                        "description": "No EOD report found for date"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter Sanctum personal access token as: Bearer <token>",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}